+++ src/Metadata/Condition/RateLimitCondition.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace JacyImp\ApiPlatformRateLimiter\Metadata\Condition;
+
+interface RateLimitCondition
+{
+}