+++ tests/Integration/Symfony/ApiPlatformRateLimiterBundleIntegrationTest.php
@@ -8,6 +8,7 @@ use ApiPlatform\Metadata\Get;
 use ApiPlatform\Metadata\Operation;
 use JacyImp\ApiPlatformRateLimiter\Exception\RateLimitExceededException;
 use JacyImp\ApiPlatformRateLimiter\Metadata\BypassRateLimit;
+use JacyImp\ApiPlatformRateLimiter\Metadata\Condition\Condition;
 use JacyImp\ApiPlatformRateLimiter\Metadata\DynamicCost;
 use JacyImp\ApiPlatformRateLimiter\Metadata\RateLimit;
 use JacyImp\ApiPlatformRateLimiter\Tests\Integration\Symfony\Fixture\FixedCostResolver;
@@ -225,7 +226,7 @@ final class ApiPlatformRateLimiterBundleIntegrationTest extends TestCase
                 RateLimit::class => new RateLimit(
                     limit: 1,
                     interval: '1 minute',
-                    when: 'test.never_apply',
+                    when: new Condition('test.never_apply'),
                 ),
             ],
         );